requirements.txt
Brotli==1.0.9
click==8.1.3
dash==2.6.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
Flask-Compress==1.13
greenlet==1.1.3
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
pandas==1.3.5
pkg-resources==0.0.0
plotly==5.10.0
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.41
tenacity==8.1.0
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.8.1
